Add function pointers + local_ethernet example
