+++ examples/local_ethernet/device/syndesi_config.h
@@ -1,9 +1,10 @@
-/* Please edit this file to configure how the syndesi library works
-*
-* Part of this file has been generated automatically by the script generate_commands.py
-* date : 22-11-01 15:11:23
+/* THIS FILE HAS BEEN AUTOMATICALLY GENERATED BY generate_commands.py
+*  
+*  ONLY EDIT PARAMETERS AS SUGGESTED.
+* date : 22-11-06 14:51:22
 */
 
+#define SYNDESI_CONFIG_INCLUDED // Do not change this
 
 /* ######## General configuration ######### */
 
@@ -16,39 +17,25 @@
 * DEVICE mode : can receive commands (like a I²C interface, SPI interface, etc...)
 */
 #define SYNDESI_DEVICE_MODE
+
 // Both can be activated at the same time to enable HYBRID mode
 
 
 /* ######## Protocols configuration ####### */
 
 /*
-* Configure what controllers are used (uncomment to enable)
+* ETHERNET
+*
+* 0 -> custom (user implemented)
+* 1 -> system
+*
 */
-
-// IP Controller
-#define USE_IP_CONTROLLER
-// Controller MODE
-// CONTROLLER_MODE_POLL : Poll for received data
-// CONTROLLER_MODE_WAIT : Wait for received data
-// CONTROLLER_MODE_IRQ : Interrupt call when data is received
-#define IP_CONTROLLER_MODE CONTROLLER_MODE_POLL
+#define SYNDESI_ETHERNET_CONTROLLER 1
 
 
-// UART Controller
-#define USE_UART_CONTROLLER
-// Controller MODE
-// CONTROLLER_MODE_POLL : Poll for received data
-// CONTROLLER_MODE_WAIT : Wait for received data
-// CONTROLLER_MODE_IRQ : Interrupt call when data is received
-//#define UART_CONTROLLER_MODE
 
-// RS485 Controller
-#define USE_RS485_CONTROLLER
-// Controller MODE
-// CONTROLLER_MODE_POLL : Poll for received data
-// CONTROLLER_MODE_WAIT : Wait for received data
-// CONTROLLER_MODE_IRQ : Interrupt call when data is received
-//#define RS485_CONTROLLER_MODE
+// Use system Ethernet
+#define USE_SYSTEM_ETHERNET
 
 
 /* ######## Callbacks configuration ######## */
@@ -72,17 +59,18 @@
 
 */
 
+
 /*
 * Request callbacks (to be used in device code)
 * Only active if DEVICE_MODE is defined
 */
-// #define USE_DEVICE_DISCOVER_REQUEST_CALLBACK
-// #define USE_REGISTER_READ_16_REQUEST_CALLBACK
+#define USE_DEVICE_DISCOVER_REQUEST_CALLBACK
+#define USE_REGISTER_READ_16_REQUEST_CALLBACK
 #define USE_REGISTER_WRITE_16_REQUEST_CALLBACK
-// #define USE_SPI_READ_WRITE_REQUEST_CALLBACK
-// #define USE_SPI_WRITE_ONLY_REQUEST_CALLBACK
-// #define USE_I2C_READ_REQUEST_CALLBACK
-// #define USE_I2C_WRITE_REQUEST_CALLBACK
+#define USE_SPI_READ_WRITE_REQUEST_CALLBACK
+#define USE_SPI_WRITE_ONLY_REQUEST_CALLBACK
+#define USE_I2C_READ_REQUEST_CALLBACK
+#define USE_I2C_WRITE_REQUEST_CALLBACK
 
 
 
@@ -90,11 +78,11 @@
 * Reply callbacks (to be used in host code)
 * Only active if HOST_MODE is defined
 */
-//#define USE_ERROR_REPLY_CALLBACK
-//#define USE_DEVICE_DISCOVER_REPLY_CALLBACK
-//#define USE_REGISTER_READ_16_REPLY_CALLBACK
-//#define USE_REGISTER_WRITE_16_REPLY_CALLBACK
-//#define USE_SPI_READ_WRITE_REPLY_CALLBACK
-//#define USE_SPI_WRITE_ONLY_REPLY_CALLBACK
-//#define USE_I2C_READ_REPLY_CALLBACK
-//#define USE_I2C_WRITE_REPLY_CALLBACK
+#define USE_ERROR_REPLY_CALLBACK
+#define USE_DEVICE_DISCOVER_REPLY_CALLBACK
+#define USE_REGISTER_READ_16_REPLY_CALLBACK
+#define USE_REGISTER_WRITE_16_REPLY_CALLBACK
+#define USE_SPI_READ_WRITE_REPLY_CALLBACK
+#define USE_SPI_WRITE_ONLY_REPLY_CALLBACK
+#define USE_I2C_READ_REPLY_CALLBACK
+#define USE_I2C_WRITE_REPLY_CALLBACK
