+++ user_config/syndesi_config.h
@@ -1,7 +1,7 @@
 /* THIS FILE HAS BEEN AUTOMATICALLY GENERATED BY generate_commands.py
 *  
 *  ONLY EDIT PARAMETERS AS SUGGESTED.
-* date : 22-11-03 16:34:10
+* date : 22-11-06 14:51:22
 */
 
 #define SYNDESI_CONFIG_INCLUDED // Do not change this
@@ -11,14 +11,14 @@
 /*
 * HOST mode : can send commands (a computer for example)
 */
-//#define SYNDESI_HOST_MODE
+#define SYNDESI_HOST_MODE
 
 /*
 * DEVICE mode : can receive commands (like a I²C interface, SPI interface, etc...)
 */
 //#define SYNDESI_DEVICE_MODE
 
-// Both can be activated at the same time to enable HYBRID mode
+// Both can be activated at the same time to enable HYBRID mode (not enabled yet)
 
 
 /* ######## Protocols configuration ####### */
@@ -64,13 +64,13 @@
 * Request callbacks (to be used in device code)
 * Only active if DEVICE_MODE is defined
 */
-//#define USE_DEVICE_DISCOVER_REQUEST_CALLBACK
-//#define USE_REGISTER_READ_16_REQUEST_CALLBACK
-//#define USE_REGISTER_WRITE_16_REQUEST_CALLBACK
-//#define USE_SPI_READ_WRITE_REQUEST_CALLBACK
-//#define USE_SPI_WRITE_ONLY_REQUEST_CALLBACK
-//#define USE_I2C_READ_REQUEST_CALLBACK
-//#define USE_I2C_WRITE_REQUEST_CALLBACK
+#define USE_DEVICE_DISCOVER_REQUEST_CALLBACK
+#define USE_REGISTER_READ_16_REQUEST_CALLBACK
+#define USE_REGISTER_WRITE_16_REQUEST_CALLBACK
+#define USE_SPI_READ_WRITE_REQUEST_CALLBACK
+#define USE_SPI_WRITE_ONLY_REQUEST_CALLBACK
+#define USE_I2C_READ_REQUEST_CALLBACK
+#define USE_I2C_WRITE_REQUEST_CALLBACK
 
 
 
@@ -78,11 +78,11 @@
 * Reply callbacks (to be used in host code)
 * Only active if HOST_MODE is defined
 */
-//#define USE_ERROR_REPLY_CALLBACK
-//#define USE_DEVICE_DISCOVER_REPLY_CALLBACK
-//#define USE_REGISTER_READ_16_REPLY_CALLBACK
-//#define USE_REGISTER_WRITE_16_REPLY_CALLBACK
-//#define USE_SPI_READ_WRITE_REPLY_CALLBACK
-//#define USE_SPI_WRITE_ONLY_REPLY_CALLBACK
-//#define USE_I2C_READ_REPLY_CALLBACK
-//#define USE_I2C_WRITE_REPLY_CALLBACK
+#define USE_ERROR_REPLY_CALLBACK
+#define USE_DEVICE_DISCOVER_REPLY_CALLBACK
+#define USE_REGISTER_READ_16_REPLY_CALLBACK
+#define USE_REGISTER_WRITE_16_REPLY_CALLBACK
+#define USE_SPI_READ_WRITE_REPLY_CALLBACK
+#define USE_SPI_WRITE_ONLY_REPLY_CALLBACK
+#define USE_I2C_READ_REPLY_CALLBACK
+#define USE_I2C_WRITE_REPLY_CALLBACK
